MainActivityをきれいに。

+++ app/src/main/java/com/example/hanah/a101bandouro/MainFragment.kt
@@ -16,6 +16,7 @@ import java.io.FileInputStream
 import java.io.FileOutputStream
 import android.content.Context
 import android.widget.Toast
+import com.example.hanah.a101bandouro.tool.DatabaseModel
 import io.reactivex.Single
 
 
@@ -23,23 +24,23 @@ import io.reactivex.Single
 /**
  * Created by hanah on 2017/11/11.
  */
-class MainFragment(val callback: Callback, context: Context) : android.support.v4.app.Fragment() {
+class MainFragment : android.support.v4.app.Fragment() {
 
     private var mediaPlayer: MediaPlayer = MediaPlayer()
     private var station = ""
-    /*private var audioTrack: AudioTrack? = null*/
-    private var contexts = context
+    private var contexts: Context? = null
 
     interface Callback {
         fun setText(station: String)
     }
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val file = NCMBFile(station + "2" + ".wav")
+    override fun onAttach(context: Context?) {
+        super.onAttach(context)
+        contexts = context
     }
 
     fun getNearStation(pointX: Double, pointY: Double, tasteful: Int) {
+        if (context == null) return
         var newStation: String
         val client = ServerClient(com.example.hanah.a101bandouro.model.Key.eki)
         client
@@ -48,15 +49,12 @@ class MainFragment(val callback: Callback, context: Context) : android.support.v
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe({
                     newStation = it.ResultSet.Point.Station.Name
-                    callback.setText(newStation)
+                    (contexts as MainActivity).setText(newStation)
                     Log.d("近くの駅", newStation + tasteful.toString())
                     if (station != newStation) {
-                        if (tasteful == 4) {
-                            playStationMusic("", tasteful)
-                        } else {
-                            station = newStation
-                            playStationMusic(station = newStation, tasteful = tasteful)
-                        }
+                        station = newStation
+                        playStationMusic(station = newStation, tasteful = tasteful)
+                        DatabaseModel(context).onDataSearch(station)
                     }
                 }, {
                     it.printStackTrace()
@@ -65,13 +63,14 @@ class MainFragment(val callback: Callback, context: Context) : android.support.v
     }
 
     private fun playStationMusic(station: String, tasteful: Int) {
+        if (context == null) return
         val file = if (station.isBlank()) {
             NCMBFile("さんぽ.mp3")
         } else {
             NCMBFile(station + tasteful.toString() + ".mp3")
         }
         file.fetchInBackground({ bytes: ByteArray?, ncmbException: NCMBException? ->
-            val tempMp3 = File.createTempFile(station + tasteful.toString() + "hogehoge", ".mp3", contexts.cacheDir)
+            val tempMp3 = File.createTempFile(station + tasteful.toString() + "hogehoge", ".mp3", contexts?.cacheDir)
             tempMp3.deleteOnExit()
             val fos = FileOutputStream(tempMp3)
             fos.write(bytes)
@@ -97,6 +96,7 @@ class MainFragment(val callback: Callback, context: Context) : android.support.v
     }
 
     fun stopMusic() {
+        if (context == null) return
         if (mediaPlayer.isPlaying) {
             mediaPlayer.pause()
             mediaPlayer.reset()
